backend/requirements.txt
Django
django-filter
djangorestframework
django-cors-headers
django-colorfield
django-extra-fields
djoser
gunicorn
psycopg2-binary
Pillow
python-dotenv